# Use mapChunks in CirceInstances.streamedJsonArray