Small clean up to use `mapChunks` instead of an implementation using `repeatPull`.