@peterneyens 2023-05-13
Finally found some time to add a benchmark.

Previous
```
[info] Benchmark                           (elems)  (elemsPerChunk)  Mode  Cnt        Score        Error  Units
[info] CirceJsonStreamBench.encode_stream      100               50  avgt    5    54294.630 ±  13951.153  ns/op
[info] CirceJsonStreamBench.encode_stream      100              500  avgt    5    50007.465 ±   2632.255  ns/op
[info] CirceJsonStreamBench.encode_stream     1000               50  avgt    5   299525.990 ±  54286.551  ns/op
[info] CirceJsonStreamBench.encode_stream     1000              500  avgt    5   251252.929 ± 101122.138  ns/op
[info] CirceJsonStreamBench.encode_stream    10000               50  avgt    5  2731966.198 ± 760253.727  ns/op
[info] CirceJsonStreamBench.encode_stream    10000              500  avgt    5  2224727.347 ± 327952.214  ns/op
```

This PR:
```
[info] Benchmark                           (elems)  (elemsPerChunk)  Mode  Cnt        Score        Error  Units
[info] CirceJsonStreamBench.encode_stream      100               50  avgt    5    52789.592 ±   7598.824  ns/op
[info] CirceJsonStreamBench.encode_stream      100              500  avgt    5    49110.386 ±   1697.552  ns/op
[info] CirceJsonStreamBench.encode_stream     1000               50  avgt    5   280540.460 ±  30941.520  ns/op
[info] CirceJsonStreamBench.encode_stream     1000              500  avgt    5   237874.710 ±   6312.693  ns/op
[info] CirceJsonStreamBench.encode_stream    10000               50  avgt    5  2564971.598 ±  92078.114  ns/op
[info] CirceJsonStreamBench.encode_stream    10000              500  avgt    5  2214598.872 ± 685056.424  ns/op
```